Correct the behaviour of loc_json to return an encoded Feature rather than an encoded Geometry.
This was being masked by OpenLayers reading the Geometry and silently
converting it to a Feature. A new test highlighted this problem and
various tests are updated. In hindsight, it seems obvious that all
the loc_json= tests use features but the loc_json method returns a
Geometry - this inconsistency is now fixed.