Reduce the bbox strategy ratio, and introduce a resFactor.
The ratio is the size of the bbox requested, compared to the size
of the map shown, and defaults to 2. Reducing this will request
smaller areas from the server and hence smaller maximum polygons,
but leads to delays when panning.

The resFactor instructs the strategy to reload if you zoom in too
far (i.e. change the resolution by a given factor). This means if
e.g. central cambridge has 50 issues, and you zoom in on Orchard
Park, after two zoom ins (resolution change 4) you'll reload just
the area that you're now looking at.