Push the thread search updates onto resque instead of processing inline.
The search updates appear to be about 10x slower on the server than
on a typical laptop, but it's not clear why. In any case, it's not
something that needs doing synchronously, so we can push this onto a
queue. Metrics suggest search updates are about 75-90% of the time taken
to post a new message to a thread.