Improve the accept invitation to joain a group screen.
This previously only had the default divise template. I've added a help message to give the user some more information, and in creating the form allowed the user to update their name and email address that they wish to use for the system as the one used may not be their preferred email address or name.

We should consider taking them to the group that they were invited to join, or highlighting that somehow.