Don't have maps nested in divs in lists. Closes #39.
We have the width of the div#map set to 100%, but IE7
couldn't figure out how wide that should be, if the *parent* div was nested
in a list container. The only hint was that it worked at all for the issues
page, which had the map correctly nested in a list item.