Rename the controllers for creating library items from messages.
Turns out that things like Message::Library::DocumentsController will
blow up in your face if there's already a Message class. Rails copes
to a certain extent, but the double nesting is too much for the tests.

Rename everything to message_library which is less that self-explanatory
but at least works.