Add groups gallery. #27
Adds a groups listing and maps. Still needs a groups in your areas tab.
There is an issue on Line 63 of the groups controller where the code will randomly not work with the
error: uninitialized constant Group::GroupsController, yet some time later with no code change, it'll
just suddenly start working again.

The javascript could potentially be simplified to reduce the duplication, though would require a bit of refactoring.