Check for name and email before inviting user.
The behaviour of mail has changed, and now throws an error if there's no To: on
the mail. Previously when submitting a blank invite form, we'd try inviting
email: "" and it would silently do nothing, and then the user would be invalid and
the form re-shown.

There's probably a better approach, but an additional benefit here is clarifying which
values we're checking, and which actions are being made, rather than just having everything
as one concatenated statement.