Patch to possibly add threading headers to cyclestreets mails
include a References header to be the same as the reply_to address, which is a valid email address for that thread. Will do basic threading.

Ideally the message ID would be populated here, and would take the message-in-thread-number that comes in far higher up in the call change, and do <msg-number>-<threadid>@domain so we get predictable msgids, and can then add the References programatically.