Support touch for the tools menu on touch screens.
Can now edit. Shows the tool menu cog all the time as many didn't realise it was there, and it couldn't be used on touch screens due to hover. Can't get rid of the menu once it appears on touch screens.  Tested on the iOS simulator only so far. It also switching to display none/block rather than the ineffient 999999px to the right as the browser has to draw a canvas of that size.