Improve profile permission redirect
If a user is not logged in and tries to view a profile that:
a) does not exist
b) is not public
they are asked to sign in.
It a user is logged in and tries to view a profile that:
a) does not exist
b) is not public and they share no groups
they see an unauthorised error.

Fixes #816