Store image heights
This way we get the layout of the page and can scroll to the correct message independant of images.
If we don't store the image heights we will scroll then as the images are loaded then the page jumps around.
See 2024ff41530273d3d3ff2f4fd53d63b9f7c5fccd for why we use window load
The db changes are to cache the image heights in the db otherwise they need to be re-calculated via
ImageMagick each time.